requirements.txt
# Django and Core Dependencies
Django>=5.0,<5.1
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Google Sheets Integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Database (optional - uncomment if using PostgreSQL)
# psycopg2-binary>=2.9.9

# Development Tools (optional)
# django-debug-toolbar>=4.2.0




